[#27155] Create AcountProvider

+++ src/main/java/com/cryptonita/app/data/providers/IAcountProvider.java
@@ -34,7 +34,7 @@ public interface IAcountProvider {
      * @param ammount the amount of the user's coins
      * @return the dto with all the information of the account
      */
-    WallerResponseDto deposit(String user, String coin, double ammount);
+    WallerResponseDto deposit(String user, String coin, long ammount);
 
     /**
      *
@@ -45,7 +45,7 @@ public interface IAcountProvider {
      * @param ammount the amount of the user's coins
      * @return the dto with all the information of the account
      */
-    WallerResponseDto withDraw(String user, String coin, double ammount);
+    WallerResponseDto withDraw(String user, String coin, long ammount);
 
     /**
      * Convenient method for clear
